feat(llm): render markdown in llmChat backlink excerpts
Rendering through the shared chat markdown pipeline lets the generic
HTML excerpt extraction handle chats too, so the dedicated chat token
walker is gone; both excerpt kinds now live in backlink_excerpts.ts.
